UVG-Agent/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
requests>=2.31
anthropic>=0.30
python-dotenv>=1.0
pydantic>=2.6